+++ packages/desktop-api/index.ts
@@ -1,0 +1,7 @@
+export * from './src/fs';
+export * from './src/dialog';
+export * from './src/window';
+export * from './src/core';
+export * from './src/event';
+export * from './src/path';
+export * from './src/shell';