+++ packages/ui-components/index.ts
@@ -1,2 +1,5 @@
 export { default as BitsyCharacter } from './src/BitsyCharacter';
 export { BitsyReel } from './src/BitsyReel';
+export { default as Header } from './src/Header';
+export { default as Sidebar } from './src/Sidebar';
+export { default as NavigationTabs } from './src/NavigationTabs';